DCOS-410 Check that the app exists before adding
It is a Marathon bug for POSTing to v2/apps to lead to an update/PUT
when the app id already exists. This has been fixed in the latest
Marathon but it causes the DCOS CLI to behave incorrectly with older
Marathon.

This adds a basic check when doing `dcos app add`. This change breaks
the existing tests that assume that `add` turns into an update. The
tests were updated to match the new semantic.