Fix dcos node ssh
Fixing Issue #291.  The old method of retrieving the Mesos master's ip
was returning the internal private IP.  We now retrieve the public IP
for the master node.  Further, ssh to slaves didn't work because we
weren't being redirected through the master as described in the code
comments.  This PR also fixes that problem.  See issue #293.

The old ssh to master test which used to pass has been removed.  This
change broke the test, but the test didn't actually prove that anything
worked.  I've filed issue #294 to fix the 'ssh' tests.

ssh now works to both master and slave nodes in DCOS clusters.