Escape newlines when rendering template
If we read an options json file which includes properly encoded newlines
like this:

    { "foo": "bar\nbaz" }

And have a template like:

    { "bzz": "{{ foo }}" }

Without this change, the resulting file would be:

    { "bzz": "bar
    baz" }

which isn't valid json.

Alternatively people would need to double escape the newline like this:

    { "foo": "bar\\nbaz" }

which is quite confusing and unexpected from a users perspective.

One thing to keep in mind is that with this change, it's not possible to
use newlines outside strings. Since that should be very uncommon and
unreadable, I don't think we need/should support that.

The proper solution to all this would be to stop using template to
render json. The simplest alternative would be to simply override
elements by merging marathon.json with some user defined json. If that
doesn't work for all use cases, we would need to come up with a json
transformator (think XSLT) or DSL.