Adding version checking to the install script.
All DCOS clusters currently download the latest `install.sh`. When we introduce backward incompatabilities, this ends up creating a less than ideal user experience. By checking the version of the cluster, we're able to install the correct CLI version for the cluster. This is a somewhat temporary fix and we'll design a more permanent one later.