inital work for SIMPLE integration of opendcre into dcos-cli - this is more as a proof of concept
to get this working, follow the developer setup instructions as laid out in the dcos-cli readme. then, simply calling the opendcre commands appropriately should do it. for now, the backend implementation is just a stub to act as a proof of concept - it hits a hardcoded url (pointing to 192.168.99.100 because that is the address of docker-machine on mac -- should update the URLs appropriatly)

the above stub simplifies things a little in that no configuration of dcos has to happen to reach opendcre (running as vaporcore southbound emulator, in docker container in the background). ideally, we will want an opendcre python client, or some client integrated w/ dcos to use the proper node that is connected to, etc.