remove v2/info call from Marathon client (#672)
We were using it to make sure we have a recent version of Marathon. Since
DC/OS CLI 0.4.0+ is only compatible with DC/OS 1.6.1, the Marathon
check is no longer necessary. It is also messing with ACLs since
v2/apps is at a user level and v2/info is an admin level.