task: update `dcos task exec` ATTACH_CONTAINER_INPUT semantics. (#856)
With the new semantics, an ATTACH_CONTAINER_INPUT call will be attempted
once before committing to a persistent connection where the actual input
is streamed to the agent.

This change is required to get around a limitation of the python client
libraries where an HTTP response is not read until after the entire
request is sent. IN our case, we *never* complete the request (until a
timeout), so we don't see the proper error unless we do things as
updated in this commit.