fix --leader flag for dcos-log
currently the command
dcos node log --leader will not get logs from mesos leader node. It will make
a request against `dcos_url` variable. We should redirect the request to a
leading mesos node unless --component dcos-marathon is used. In this case
we have to redirect the request to a leading marathon master (which can be
a totally different node)