Only set tty into raw mode for `dcos task exec` when interactive.
Previously, the terminal for 'dcos task exec' was set into raw mode
anytime the '--tty' flags was set (even if '--interactive' was not set).
This would cause the CLI to hang when '--interactive' was not set since
control characters are not interpreted by the local tty once it is set
into raw mode (they are sent to STDIN instead). Since we don't actually
stream STDIN to the remote container without '--interactive' , its safe
to leave the tty in default mode and allow the user to '^C' out of it.

This commit addresses this issue.