Updated attach input path of 'dcos task exec'.
Previously, we would error out no matter what type of exception was
thrown while trying to establish an ATTACH_CONTAINER_INPUT session with
an agent. This would cause us to sometimes drop the output of command we
launched because it happened to complete before we made the
ATTACH_CONTAINER_INPUT call (because the command actually didn't need to
accept input and just ran to completion). We should only error
out if we we are certain that the error occurred while the container was
still running.

This commit adds logic to update this error handling properly.