dcos node ssh --private-ip=<private-ip>
People working on the soak clusters at Mesosphere receive alerts when
something unexpected is happening, e.g. 'Some containers failed to
launch on an agent on agent:<private_ip>'.

This new command allows developers to quickly take a look without
having to check the Mesos UI to find the Mesos ID of an agent. Another
use case is when a developer sees an unhealthy AWS instance that is a
DC/OS private agent without a public IP.