dcos node: Add --field option
This lets the user control what fields are printed in the table and
allows printing arbitary data that wasn't possible to show before,
except when using `--json`.

Here's an example of referencing a custom attribute called `ethos_role`:

```
$ dcos node --field=ip --field=id --field=attributes.ethos_role
     IP                          ID                    ATTRIBUTES.ETHOS_ROLE
10.10.3.201   74df965e-a46e-46a9-8d2a-639db0af0e25-S3          public
10.10.12.29   74df965e-a46e-46a9-8d2a-639db0af0e25-S2         control
10.10.12.174  74df965e-a46e-46a9-8d2a-639db0af0e25-S0          worker
10.10.12.237  74df965e-a46e-46a9-8d2a-639db0af0e25-S1          proxy
```

By using a `:` in a field name, you can give it a custom heading in the
table.

$ dcos node --field=ip --field=id --field=ethos_role:attributes.ethos_role
     IP                          ID                    ETHOS_ROLE
10.10.3.201   74df965e-a46e-46a9-8d2a-639db0af0e25-S3    public
10.10.12.29   74df965e-a46e-46a9-8d2a-639db0af0e25-S2   control
10.10.12.174  74df965e-a46e-46a9-8d2a-639db0af0e25-S0    worker
10.10.12.237  74df965e-a46e-46a9-8d2a-639db0af0e25-S1    proxy
```

Note that the table is always sorted by the first field specified and if
it's `ip`, it sorts numerically instead of alphabetically.

One more example:

```
$ dcos node --field=disk\ used:used_resources.disk --field=endpoint:pid --field=version
DISK USED           ENDPOINT           VERSION
   0.0     slave(1)@10.10.3.201:5051    1.0.1
  4096.0   slave(1)@10.10.12.174:5051   1.0.1
  4096.0   slave(1)@10.10.12.237:5051   1.0.1
  5096.0   slave(1)@10.10.12.29:5051    1.0.1
```