Fixed bug returning the package JSON for 'CosmosPackageVersion' class.
The JSON representation of the CosmosPackageVersion recently changed to
wrap the contents of the previous package JSON in a 'package' key (so
that other fields can be added at the top level in the future). In order
to get at the actual package now, we need to index into the 'package'
key and return it.