Updated consolidated tox.ini files and made them workable on windows.
Some extra environment variables are necessary on windows to get the CLI
to pass the integration tests when using 'tox'. Specifically, the HOME*
environment variables are needed to allow python's 'expanduser()'
function to succeed when resolving the '~' path for the default
DCOS_DIR. Additionally, the OPENSSL_CONF variable is need to allow SSL
connections to be made properly using the windows OpenSSL library.
Without these additional variables things seem to die in unexpected
ways.