Updated line count in 'test_ls_long' for '*.logrotate.state' files.
These files used to be present when running this test, but now they are
not. I wasn't able to find the commit that changed things so that these
files don't appear anymore, but this test is now consistently failing
since they are not present. This commit reduces the number of lines
expected in the output since these files are no longer present.