Updated 'test_ls_long()' to look for a range of lines.
Previously, this test was flaky because sometime it would come up with 5
lines and other times 7. This flakiness is due to some idiosyncrasies
with 'logrotate' where sometimes it creates a '*.logrotate.state' file
during the test run, and other times it doesn't. When it does, there
will be 7 lines in the output. When it doesn't, there will be 5.