remove order dependant tests
test_install_with_id, test_uninstall_with_id and test_uninstall_all are
a combination of tests which passes only as long as they are run
sequentially.

As passing the --app-id for installing a package is not supported
anymore, it doesn't make much sense to keep them. There are other tests
covering this.