Updated Jenkinsfiles to run with CLI_TEST_MASTER_PROXY=true.
This setting forces the `--master-proxy` flag to be set when running the
various `dcos node ssh` commands during integration testing. Without
this flag, `ssh` commands ran outside of a DC/OS cluster will never
succeed in connecting because they all try to connect through a private
IP address. With this flag, they all get proxied through the public IP
of a master node.

It is a weird quirk in the testing infrastructure that any `dcos node
ssh` tests were ever passing. Right now, the tests simply check to see
if a connection has been *attempted* rather than actually checking to
make sure it has *succeeded*.

This commit makes sure we actually carry out the connection
successfully. A subsequent commit will fix the testing infrastructure to
ensure that we fail if the connection is never established.