Updated `node_ssh_output()` to fail if ssh connection never established.
Previously, this command simply issued an ssh command, waited 5 seconds
to ensure the command was started successfully, and then killed it. It
made no attempt to actually ensure that the ssh connection was
established successfully.

This commits puts a *temporary* fix in place to make sure that the
connection is actually established before terminating the ssh
connection. We need a more robust solution that actually waits on the
connection with a timeout (instead of unconditionally waiting and just
killing the process when it doesn't exit). For now this will work,
however, until we have time to put a more robust solution in place.