Add a command to mark an agent as gone
This adds a `dcos node decommission` command which takes as argument an
agent id. It will then mark it as gone using the Mesos operator API.

Currently the integration test is skipped because it requires the
opposite of that action in order to put the cluster in a proper state
again after that test.

cf. https://jira.mesosphere.com/browse/DCOS-18981