Remove the hardcoded 1 second timeout for some HTTP calls
This might be too short in case of network latency or if for any reason
the cluster is not very responsive but still works fine. In such cases
commands like "cluster setup" or "cluster --version" might fail.

There is already a 5 seconds timeout by default that we can rely on.

https://jira.mesosphere.com/browse/DCOS_OSS-1763