Catch other exceptions than SSLError in _needs_cluster_cert
Otherwise in case of an unreachable cluster URL there is a stack trace
shown-up. This is the first request going out so it should catch other
exceptions too.

Tests have been added to make sure the CLI doesn't hang for too long or
print unexpected errors in case of unreachable URLs.