Lookup for DCOS_CLUSTER to determine if a cluster is attached
We support the DCOS_CLUSTER env variable since #994, we should take it
into account when checking if a cluster is attached.

This is useful for automated scripts where one would want to attach to a
cluster without affecting potential concurrent users (because of the
shared attached file).

cf. https://github.com/dcos/dcos-cli/issues/595