HTTP timeout improvements
This leverages the tuple timeout of the requests package in order to
distinguish between the connection timeout and the response read
timeout.

When a timeout is passed as a numeric value, it refers to the
read_timeout.

When no explicit timeout is passed, the `core.timeout` value in the
config file is now used, or if not present, defaults to 180. This also
refers to the read_timeout.

The connection timeout defaults to 5. In order to overwrite it, one must
pass a tuple of the form `(connection_timeout, read_timeout)`.

Fixes :

https://jira.mesosphere.com/browse/DCOS_OSS-1956
https://jira.mesosphere.com/browse/DCOS_OSS-1957