Pass the config when checking if a request is to DC/OS
The toml_config argument can be used to specify a non default config to
use when making a request. This is useful when making a request to a
cluster which is not currently attached.

In such cases the config from the given cluster should be used to
determine if it's an external request or not.