Consider that dcos version metadata might be invalid
I had an issue I cannot reproduce where the JSON returned by this
endpoint was invalid, probably because of a wrongly installed /
unhealthy cluster.

In such case it becomes impossible to run `dcos cluster list` for
example as an exception is thrown, this change makes the code more
resilient to this kind of problem.