Wait for an app count after uninstalling a package
Waiting for marathon deployments to finish is not enough, the marathon
app might still be running for some time afterwards.

As we have many tests using different packages, it seems unrealistic to
check for marathon apps directly.

This introduces a `count_apps()` helper function. Before installing a
package, the number of apps is stored. On uninstall, it then waits for
the original number of apps before moving-on.