Bootstrap the Go DC/OS CLI.
The Go DC/OS CLI is a stripped-down version of the current CLI acting
like a wrapper. It only contains the `config`, `auth`, and `cluster`
subcommands. Other subcommands will get installed automatically in the
background while setting-up a cluster.

This wrapper allows users to have only one DC/OS CLI that works with
different versions of DC/OS clusters, the correct CLI to communicate
with the cluster being downloaded and selected automatically by the
wrapper. This will allow the distribution of the Go CLI on standard
channels such as Homebrew in the future.

This patch bootstraps the Go project with the base CLI structure and
the config `set`, `unset`, and `show` subcommands. In order to do so,
it introduces a config package. Its technical overview can be found
in the `design/config.md` file.