Update package integration tests for DC/OS 1.12
This uses the helloworld universe image generated by the following PR :

https://github.com/mesosphere/universe/pull/1751

It also removes the frozen version of the main Universe. The reason
for this was probably to satisfy all fixtures from package integration
tests. But this introduces technical debt as each time we need to
upgrade it there is a bunch of tests failing and we have to update all
sort of messages assertions which comes from Cassandra, Chronos and
other packages. This updates some of the tests to use helloworld
instead.