Fixed pattern matching for task IDs in 'task exec'.
Previously, there were problems with the way patterns were matched for
the task ID when calling `task exec`. Specifically, it didn't follow the
semantics indicated in the help string for (1) matching regular
expressions and (2) matching any substring within the ID, not just the
prefix of the ID.

The reason we did not adhere to these semantics is because we were
trying to match the exact semantics of `docker exec`, which only allows
prefix matching, not arbitrary regex or 'contains()' like semantics for
pattern matching the ID.

With this change, we have decided to break from the docker semantics and
use the default pattern matching semantics provided by the dcos-cli
library command for pattern matching task ids.

Unfortunately, there is a bug in this library, and regular expressions
are not supported! However, arbitrary substring pattern mathing is.

This commit, therefore, updates the semantics of `task exec` to adhere
to 'contains()' like smeantics within the task ID for the 'task_id'
parameter passed to it, but does not yet support proper regular
expression pattern matching.

A subsequent commit will be added to fix regex matching in library call
implementing task ID pattern matching.