Introduce a ConfigManager
The config manager acts as a repository for configurations. It is
currently only able to find configs but in the future it will be
responsible for creating and removing configs as well.

This commit also adds several test fixtures to make sure the config
manager behaves correctly under different scenarios and edge-cases.