Introduce a CLI context
Introduce a Context struct. It is created based on various
abstractions (eg. io.Reader for stdout/stderr) or by calling
DefaultContext() which will be backed by the OS.

Once created, the context must be passed as an argument to each subcommand
constructor and they must use it to interact with the environment (print
output, read env vars or files, etc.) or instanciate
environment-dependent structs from other packages.