Config refactoring (#1193)
* Update config design page

This refines the goal of the package and adjusts the implementation
plan accordingly.

* Refactor the config package

The Config struct becomes Cluster and is moved into a `core` subpackage.
Decoupling it from the main package makes the config agnostic from any
domain model, rather, Cluster is just an example of how one can use a
Config to fits their application needs.

The marshalling/unmarshalling logic has been removed in favor of a
lighter proxy object (Cluster) which gets/sets values in the Store on
the fly. Also, only relevant `core` and `cluster` fields now have
getters/setters.

* Introduce a ConfigManager

The config manager acts as a repository for configurations. It is
currently only able to find configs but in the future it will be
responsible for creating and removing configs as well.

This commit also adds several test fixtures to make sure the config
manager behaves correctly under different scenarios and edge-cases.

* Introduce a CLI context

Introduce a Context struct. It is created based on various
abstractions (eg. io.Reader for stdout/stderr) or by calling
DefaultContext() which will be backed by the OS.

Once created, the context must be passed as an argument to each subcommand
constructor and they must use it to interact with the environment (print
output, read env vars or files, etc.) or instanciate
environment-dependent structs from other packages.

* Move Cluster out of the config package

This makes it clearer that the config package is independent of it and
it is just an example of how one can use the config package.