Create a custom map type for login providers
As Go randomizes iteration over maps, the `dcos auth list-providers`
currently lists providers in an unstable order. This can be confusing
and error-prone for users in some cases (eg. if we are printing a list
to select from).

This introduces a custom type with a `Slice()` method returning the
providers in a constant order (sorted by ID).