Introduce a login flow
It allows to initiate a login flow from a terminal, based on login flags
and an HTTP client.

The flow has to first select a login provider, either explicitly,
implicitly, or manually. It then triggers its client-method. Finally,
the ACS token returned by DC/OS is stored in the current config file.