Implement `dcos auth login` (#1231)
* Add jwt-go as a dependency

It will be used to generate a service login token from a service account
provate key.

* Add login flags

Login flags are represented in a struct. They can be resolved, meaning
that the files they point to will be read (--password-file,
--private-key), the --password-env flag would also be looked-up from the
environment. Based on a resolved flag set, one can also check whether or
not there is an implicit login provider.

* Introduce a login flow

It allows to initiate a login flow from a terminal, based on login flags
and an HTTP client.

The flow has to first select a login provider, either explicitly,
implicitly, or manually. It then triggers its client-method. Finally,
the ACS token returned by DC/OS is stored in the current config file.

* Add a `dcos auth login` command

This commands relies on the login flow and flags. There is also a helper
method in the context to initiate a flow and start it.