Refactor httpclient options
It introduces a new options struct to share the same type for client
and request options.

It gives more flexibility as it allows for example to set the ACS token
or the logger on a per-request basis. It also reduces name collision
(eg. Timeout and RequestTimeout).

It also adds a NoFollow option for HTTP redirects, passing this option
when creating the client will prevent it from following HTTP redirect
responses.