Move back cli.Cluster to config.Cluster
As it's quite used across the project, having it in the cli package
doesn't give the flexibility to break down other things into their own
pacakge. This was for example the case for the setup package, as it'd
depend on the cli package for the Cluster model, and the cli package
would have a circular dependency with it.