Set the logging level outside of cobra
Having it in the pre run hook discards all the logs from the plugin
manager, this is because the plugin manager is invoked while creating
the cobra command tree. The hook is only invoked later while actually
calling a command.

Declaring the flag outside of cobra and setting the log level directly
fixes this problem.

https://jira.mesosphere.com/browse/DCOS_OSS-3822