Setup integration tests
This adds integration tests for the new Go CLI. They are run from
Jenkins and written in Python, they rely on pytest, and dcos-e2e to spin-up
the DC/OS cluster.

As the test suite doesn't alter the cluster state, a single cluster is
shared between all platform tests (Linux, macOS and Windows) which all
run in parallel.

The main goal here is not to have full test coverage, but to setup the
base structure for integration tests.

https://jira.mesosphere.com/browse/DCOS_OSS-3915