Pass the DCOS_CLI_EXECUTABLE_PATH env var to plugins
This aligns the plugin invocation with the design document.

This environment variable can be used by plugins to invoke the main CLI.
This is useful to get config values for example.

There is a second change which sets the path install option when installing
a filesystem plugin, without it local plugins can't be installed.

https://jira.mesosphere.com/browse/DCOS_OSS-3981