Intercept "dcos package install dcos-core-cli"
This is a temporary fix in place for the core plugin not being able
to update itself. In the long-term we should come-up with a system
which is able to update running binary executables.

For now we'll intercept the `dcos package install dcos-core-cli` and
trigger an installation from the top-level CLI.

https://jira.mesosphere.com/browse/DCOS_OSS-3985
https://unix.stackexchange.com/questions/138214/how-is-it-possible-to-do-a-live-update-while-a-program-is-running#answer-138241