Accept default values on confirm prompts
This adds a second argument to the Confirm() function which acts as a
fallback when users only perss [ENTER]. It is a common practice to have
prompts like:

    Please confirm [Y/n]:

In such cases the capital Y indicates that it is the default action.

This also switches the fingerprint prompt to not have a default value,
as it is contains sensitive information which should explicitly be
approved by the user.

https://jira.mesosphere.com/browse/DCOS_OSS-4030