Support DCOS_ACS_TOKEN when setting-up a cluster
DCOS_ACS_TOKEN is supported in most commands except cluster setup.

This updates the setup command to skip login and use the ACS token from
this env var when it exists. This simplifies other teams scripts as they
sometimes get the ACS token by their own mean and expect the CLI to be
configured with it.

On the other hand we didn't want to add an "--acs-token" flag to the
command as we don't encourage this behaviour.

https://jira.mesosphere.com/browse/DCOS-41803