Support browser-prompt-authtoken login client method
It was mistaken with browser-prompt-oidcidtoken-get-authtoken,
which requires the user to continue the flow in the browser
and then paste a login token to the terminal, as opposed to
an ACS token.

This commit clarifies the login design overview by adding that 5th login
client method and updates the code accordingly.

This was discovered while working on DCOS-41848.

https://jira.mesosphere.com/browse/DCOS_OSS-4134