Pass additional env vars to plugins
This is a design update to the plugin system in order to pass
additionnal env variables to plugins whenever a cluster is configured.

The main goal being to make it easier for plugins to interact with
DC/OS, as they currently have to read the config TOML file by themselves
or invoke back the `dcos` binary.

https://jira.mesosphere.com/browse/DCOS_OSS-4101