Pass cluster specific environment variables to plugins
When a plugin is invoked with a cluster currently attached, a given set
of environment variables are passed.

- DCOS_URL: The base URL of the DC/OS cluster without a trailing slash
(eg. https://dcos.example.com).

- DCOS_ACS_TOKEN: The authentication token for the current user.

- DCOS_TLS_INSECURE: Indicates whether the CLI is configured with an
insecure TLS setup. This is set to 1 when the DCOS_URL scheme is http://
or core.ssl_verify is set to false.

- DCOS_TLS_CA_PATH: Unless DCOS_TLS_INSECURE=1, this env var indicates the
path to the CA bundle to verify server certificates against. When it is
not set, the system's CA bundle must be used.

https://jira.mesosphere.com/browse/DCOS_OSS-4101