Make browser-prompt_authtoken check for 200/403 status codes
Follow-up from https://github.com/dcos/dcos-cli/pull/1338

This changes the browser-prompt_authtoken client method to expect a 200
or 403 status code in case of a valid token instead of accepting
anything else than a 401. Otherwise it should also have been checking
for 50X status codes.

This is the same behavior as the Python CLI:

https://github.com/dcos/dcos-core-cli/blob/8a51a0d05ad3ece50d1041bede5a7638f3ed13a6/python/lib/dcos/dcos/auth.py#L295