Auto-install plugins based on an experimental env var
This impletements the behavior designed in the following commit:

cdfb7401894775ba6efc8ef68ff9ca26562e5f41

It removes the `--no-plugin` option if favor of an experimental env var
called `DCOS_CLI_EXPERIMENTAL_AUTOINSTALL_PLUGINS`.

When this var is set, the CLI will autoinstall the default plugins (core
and, if applicable, enterprise).

As it's an experimental feature we're not trying to handle errors
gracefully, any error in that step will lead to a non-zero exit code
(even if it's only related to the enterprise plugin).

https://jira.mesosphere.com/browse/DCOS_OSS-4186