[design] Plugin auto-install proposal
The goal of this design proposal is to improve the dcos-core-cli and
dcos-enterprise-cli auto-installation in order to be able to remove
the core CLI bundling and make the auto-installation a non-experimental
feature.

It relies on a fallback mechanism to support most of the cases,
before eventually failing with a link to a web-page where the user can
manually get the download URLs.