Added an 'index.html' file to represent the index of all CLI artifacts.
For now, this file is unusable on it's own. In a subsequent commit, we
will add all of the assets necessary for this file to render properly,
as well as a set of scripts to upload this file (and all of its assests)
to S3 so that end users can see it.