Add support for relative DCOS_DIR
Prior to this patch, setting a relative DCOS_DIR would cause the CLI to
fail when invoking a plugin starting from the second invocation. It
happened because the CLI prepends the DC/OS dir to a command path when
it sees it as a relative path. When a DC/OS dir is relative, that
operation would happen at every plugin invocation and grow the path
indefinitely.

This updates the DCOSDir method to prepend the current directory in case
DCOS_DIR is a relative path. It also makes it return an explicit error
in case anything goes wrong.

https://jira.mesosphere.com/browse/DCOS_OSS-4405