Updated login flow to fix provider selection when uid/password given.
This resolves two issues seen in https://jira.mesosphere.com/browse/DCOS-45239.

First, if a user provides a uid/password, we default to the provider type
'DCOSUIDPassword'. This makes the behavior of the CLI the same as before.

Second, we have updated the error message when our prompt does not work
due to an input that we cannot parse.