Updated login flow to detect when dcos-users provider is available.
In a previous commit, we used the provider ID to detect a specific
provider  available for a cluster. This was wrong, the value
'DCOSUIDPassword' we are looking for being the provider type.

We should have caught the issue with an integration test but we
haven't added one due to the wish to fix the DC/OS EE CI ASAP.

Co-authored-by: Armand Grillet <armand.grillet@outlook.com>
Co-authored-by: Bilal Amarni <bilal.amarni@gmail.com>