Added support for 'providers' endpoint in open-source clusters.
With DC/OS 1.13, open source DC/OS now provides the endpoint
'/acs/api/v1/auth/providers'. This commit fixes our login
flow that was previously using that endpoint to
differentiate the type of clusters to log in.

We also have updated two tests as we now always get an
'authHeader' when getting the providers of a clusters.