Add a script verifying CLI artifacts
This adds a script which verifies that CLI artifacts are properly
deployed.

It scans the tags in dcos/dcos-cli and gets the latest one for each
minor version: 0.5.x, 0.6.x, and 0.7.x. It then downloads CLI artifacts
for each supported platform / DC/OS version and runs `dcos --version` on
them. The output is parsed and the script fails if `dcoscli.version`
contains an unexpected value.

It also checks the testing endpoint, which should have as a version the
latest commit of the master branch.

The "latest" endpoints are commented for now because they are not live
yet (it should happen on the next 0.7.x release).

The build can be manually run from here :

https://jenkins.mesosphere.com/service/jenkins/job/public-dcos-cluster-ops/job/mesosphere-dcos-cli/job/verify-artifacts/