Return plugin after installation
This returns the installed plugin when the Install() method is called.
This is useful in some cases where the caller needs to do additional
work with a plugin once it's installed, eg. print the new subcommands.

This also updates the `dcos plugin add` command to print the new
commands whenever a plugin is installed.