Pass plugin config as environment variables
This makes it possible for a plugin to retrieve its config values
through environment variables when it's invoked.

For example, assuming a command `foo` in a given plugin. When the
configuration file contains the `bar` config for the `foo` section,
it will be passed to it as the `DCOS_FOO_BAR` environment variable.