Pass plugin config as environment variables (#1499)
This makes it possible for a plugin to retrieve its config values
through environment variables when it's invoked.

For example, assuming a command `foo` in a given plugin. When the
configuration file contains the `bar` config for the `foo` section,
it will be passed to it as the `DCOS_FOO_BAR` environment variable.

The CLI version is now passed to plugins during invocation, as a
`DCOS_CLI_VERSION` environment variable.

We should design a way for plugins to expose the config keys they accept
in their plugin.toml file and the CLI would then be able to retrieve
these keys dynamically at runtime.