Discard the local login server on non-desktop environment
For the `dcos-oidc-auth0` login provider, when the user operates the CLI
from a container or a remote machine, the local webserver will not run
on the same machine as the browser. Later on the browser will try to
send the token to eg. `localhost:43383`, which won't be reachable.

When this situation arises, we had designed the UI to fallback to
printing the login token for copy-pasting, but once the CLI has
successfully initiated the local web server, it doesn't accept input
from STDIN at all, so the user has no chance to copy-paste the token.

This commit updates the CLI to fallback to reading from STDIN when the
browser can't be opened.

Note: as an alternative, it would have been possible to concurrently
wait for data on both the local web server and STDIN, but the current
fix is simpler and covers the container / remote use-cases as expected.

https://jira.mesosphere.com/browse/DCOS_OSS-5591