Do not comsume all input data on prompt
As the `prompt.Input()` function relies on a `bufio.Scanner`, it
currently scans all the lines and would only return the first one.

As the scanner is not re-used across multiple calls, this causes input
data to be discarded for later calls when there is multiline input.

As a concrete example, the following currently wouldn't work:

    echo "username\npassword" | dcos auth login --provider=dcos-users

This is however not a common case as users would usually rather do:

    dcos auth login --username=username --password=password