Check TLS config when creating HTTP Client (#1513)
* Check TLS config when creating HTTP Client

We do not check if TLS is set up properly. When we can't read
certificate or it was invalid we silently swtiched to insecure mode.
This PR changes this behaviour. Whenever HTTP Client is created it
checks for errors and returns one if necessary. The only exception are
`list` and `setup` commands. List will allow user to list all clusters
but call to attach will fail. `setup` uses `insecure`HttpClient` by
default – to download certs.

Fixes: https://jira.mesosphere.com/browse/DCOS-61005